third_party/blink/web_tests/fast/dom/Range/range-created-during-remove-children-expected.txt

PASS ranges["blur"].startOffset is 0
PASS ranges["blur"].endOffset is 0
PASS ranges["DOMNodeRemovedFromDocument"].startOffset is 0
PASS ranges["DOMNodeRemovedFromDocument"].endOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
